Move tests configuration into conftest.py
